Refactor: Utils.tsx Location Change
